# zad4_project/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1",
    "title": "Program obsługi meczy - dokumentacja Swagger",
    "description": "Dokumentacja REST API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/country": {
      "get": {
        "tags": [
          "Country"
        ],
        "description": "Pobiera wszystkie kraje",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Country"
        ],
        "description": "Wstawia kraj",
        "parameters": [
          {
            "name": "body",
            "description": "dane potrzebne do utworzenia kraju",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "name"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/country/{countryId}": {
      "get": {
        "tags": [
          "Country"
        ],
        "description": "Pobiera mecz o wskazanym id",
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID kraju",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/match": {
      "get": {
        "tags": [
          "Match"
        ],
        "description": "Pobiera wszystkie mecze lub mecze rozegrane przez podany kraj",
        "parameters": [
          {
            "name": "countryId",
            "in": "query",
            "description": "ID kraju (opcjonalne)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Match"
        ],
        "description": "Wstawia mecz",
        "parameters": [
          {
            "name": "body",
            "description": "dane potrzebne do utworzenia Meczu",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dateOfMatch": {
                  "type": "string",
                  "example": "2000-01-01"
                },
                "country1Id": {
                  "type": "string",
                  "example": ""
                },
                "country2Id": {
                  "type": "string",
                  "example": ""
                },
                "pointsCountry1": {
                  "type": "number",
                  "example": 0
                },
                "pointsCountry2": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/match/{matchId}": {
      "get": {
        "tags": [
          "Match"
        ],
        "description": "Pobiera mecz o wskazanym id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID meczu",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Match"
        ],
        "description": "Usuwa mecz o wskazanym id",
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID meczu"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}